remove redundant cast to int

+++ src/Utils/FileUtils.cpp
@@ -641,7 +641,7 @@ std::vector<std::string> FileUtils::findFileNamesByWildcard(const std::string& p
 std::string FileUtils::calcHash(const std::string& content) 
 {
   QCryptographicHash hash(QCryptographicHash::Md5);
-  hash.addData(QByteArray::fromRawData(content.data(), static_cast<int>(content.size())));
+  hash.addData(QByteArray::fromRawData(content.data(), content.size()));
   QByteArray hexResult = hash.result().toHex();
   return std::string(hexResult.constData(), hexResult.size());
 }
@@ -662,7 +662,7 @@ std::string FileUtils::calcFileContentHash(const std::filesystem::path& filePath
   while (!file.atEnd()) {
     qint64 bytesRead = file.read(buffer.data(), bufferSize);
     if (bytesRead > 0) {
-      hash.addData(QByteArray::fromRawData(buffer.constData(), static_cast<int>(bytesRead)));
+      hash.addData(QByteArray::fromRawData(buffer.constData(), bytesRead));
     } else {
       break; // read error or EOF
     }
